refactor, test prep
