+++ Gruntfile.js
@@ -1,10 +1,22 @@
 module.exports = function(grunt) {
   'use strict';
+  require('load-grunt-tasks')(grunt);
   grunt.initConfig({
     eslint: {
-      src: ["lib/**/*.js"]
+      options: {
+        configFile: '.eslintrc'
+      },
+      target: ["lib/**/*.js"]
+    },
+    mochify: {
+      options: {
+        reporter: 'spec'
+      },
+      unit: {
+        src: ['lib/angular/*-spec.js']
+      }
     }
   });
-  grunt.loadNpmTasks("grunt-eslint");
-  grunt.registerTask('default', ['eslint']);
+  grunt.registerTask('eslint', ['eslint']);
+  grunt.registerTask('test', ['mochify:unit']);
 };
