+++ lib/angular/directive-spec.js
@@ -1,0 +1,12 @@
+var angular = require('angular');
+var should = require('should');
+var sinon = require('sinon');
+require('angular-mocks');
+
+var initModule = require('../../test/initModule');
+
+describe('File file-detail Directive'), function() {
+  before(function() {
+    initModule();
+  })
+};
