+++ lib/angular/directives.js
@@ -1,0 +1,10 @@
+'use strict';
+
+
+angular.module('wfm.file.directives', ['wfm.core.mediator']);
+
+module.exports = 'wfm.file.directives';
+require('./directives/wfmImgDirective');
+require('./directives/fileDetailDirective');
+require('../../dist');
+
