+++ package.json
@@ -10,7 +10,7 @@
   "scripts": {
     "build": "wfm-template-build -m 'wfm.file.directives'",
     "watch": "wfm-template-build -w -m 'wfm.file.directives'",
-    "test": "grunt eslint"
+    "test": "grunt test"
   },
   "keywords": [
     "wfm",
@@ -21,6 +21,7 @@
   "dependencies": {
     "base64-stream": "0.1.3",
     "express": "4.14.0",
+    "fh-wfm-mediator": "0.0.15",
     "lodash": "4.7.0",
     "multer": "1.1.0",
     "q": "1.4.1",
@@ -28,8 +29,17 @@
     "uuid-js": "0.7.5"
   },
   "devDependencies": {
-    "grunt": "^1.0.1",
-    "grunt-eslint": "^18.0.0",
-    "fh-wfm-template-build": "0.0.9"
+    "angular": "^1.5.9",
+    "angular-mocks": "^1.5.9",
+    "fh-wfm-template-build": "0.0.9",
+    "grunt": "0.4.5",
+    "grunt-eslint": "18.1.0",
+    "grunt-mochify": "^0.3.0",
+    "load-grunt-tasks": "^3.5.2",
+    "mochify": "^2.18.1",
+    "proxyquireify": "^3.2.1",
+    "should": "^11.1.1",
+    "sinon": "^1.17.6",
+    "sinon-as-promised": "^4.0.2"
   }
 }
