+++ test/initModule.js
@@ -1,0 +1,10 @@
+var angular = require('angular');
+
+
+module.exports = function initModule() {
+  try {
+    angular.module('wfm.file.service');
+  } catch (e) {
+    angular.module('wfm.file.service', []);
+  }
+};
